# perf(init): Use random for session-key in fish