--- a/src/init/starship.fish
+++ b/src/init/starship.fish
@@ -18,4 +18,4 @@ function fish_mode_prompt; end
 export STARSHIP_SHELL="fish"
 
 # Set up the session key that will be used to store logs
-export STARSHIP_SESSION_KEY=(::STARSHIP:: session)
+export STARSHIP_SESSION_KEY=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)