@matchai 2020-11-21
If the performance payoff is within a margin of 1-2ms, I'm not sure it's worth complicating the setup scripts further.