@Glandos 2020-11-21
You're right, it's not the goal of performance, but more about simplicity. So I updated the code.

I also like the idea of avoiding forking subshell for simple tasks.