@davidkna 2020-11-21
I also think given the mixed speed improvement and that `starshp session` generates better session keys that the old `starship session` is preferable.
I still think it might be worth to do some mac and windows (does fish support windows?) benchmarks because iirc they start binaries more slowly than linux.