@rouge8 2021-01-02
I ran the benchmarks above on macOS 10.15.7 on a 2020 Intel MacBook Pro  and the speed improvement is more significant here -- 55% faster!:

```console
❯ hyperfine "fish -c 'export X=(starship session)'" "fish -c 'export X=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)'" "fish -c 'export X=(random 1000 9999)(random 1000 9999)(random 1000 9999)(random 1000 9999)'"
Benchmark #1: fish -c 'export X=(starship session)'
  Time (mean ± σ):      18.8 ms ±   0.7 ms    [User: 11.5 ms, System: 5.9 ms]
  Range (min … max):    16.8 ms …  20.4 ms    139 runs
 
Benchmark #2: fish -c 'export X=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)'
  Time (mean ± σ):      12.2 ms ±   0.6 ms    [User: 8.2 ms, System: 3.6 ms]
  Range (min … max):    11.2 ms …  17.1 ms    171 runs
 
Benchmark #3: fish -c 'export X=(random 1000 9999)(random 1000 9999)(random 1000 9999)(random 1000 9999)'
  Time (mean ± σ):      12.3 ms ±   0.7 ms    [User: 8.3 ms, System: 3.6 ms]
  Range (min … max):    11.0 ms …  16.0 ms    198 runs
 
Summary
  'fish -c 'export X=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)'' ran
    1.01 ± 0.08 times faster than 'fish -c 'export X=(random 1000 9999)(random 1000 9999)(random 1000 9999)(random 1000 9999)''
    1.55 ± 0.10 times faster than 'fish -c 'export X=(starship session)''
```