@matchai 2021-01-02
Running the same test, I am seeing similar results on a 2018 Macbook Pro.
(FWIW, setting `$HOME` to a tempdir stops fish from resolving to your personal config in `~/.config/fish`)

```
❯ env HOME=/tmp hyperfine "fish -c 'export X=(starship session)'" "fish -c 'export X=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)'" "fish -c 'export X=(random 1000 9999)(random 1000 9999)(random 1000 9999)(random 1000 9999)'"
Benchmark #1: fish -c 'export X=(starship session)'
  Time (mean ± σ):      14.5 ms ±   0.6 ms    [User: 7.6 ms, System: 5.5 ms]
  Range (min … max):    13.7 ms …  17.0 ms    139 runs

Benchmark #2: fish -c 'export X=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)'
  Time (mean ± σ):       8.4 ms ±   0.5 ms    [User: 4.9 ms, System: 3.2 ms]
  Range (min … max):     7.5 ms …  10.4 ms    232 runs

Benchmark #3: fish -c 'export X=(random 1000 9999)(random 1000 9999)(random 1000 9999)(random 1000 9999)'
  Time (mean ± σ):       7.9 ms ±   0.5 ms    [User: 4.7 ms, System: 2.9 ms]
  Range (min … max):     7.0 ms …  11.0 ms    248 runs

Summary
  'fish -c 'export X=(random 1000 9999)(random 1000 9999)(random 1000 9999)(random 1000 9999)'' ran
    1.06 ± 0.09 times faster than 'fish -c 'export X=(string sub -s1 -l16 (random)(random)(random)(random)(random)0000000000000000)''
    1.83 ± 0.14 times faster than 'fish -c 'export X=(starship session)''
```